Переносим функции для рендеринга из `utils.js` в `utils/render.js` и переносим оставшийся файл `utils.js` в `utils/common.js`.
