+++ src/components/task-edit.js
@@ -1,6 +1,6 @@
-import {MONTH_NAMES, DAYS, COLORS} from "../const.js";
-import {formatTime} from "../utils.js";
 import {AbstractComponent} from "./abstract-component.js";
+import {MONTH_NAMES, DAYS, COLORS} from "../const.js";
+import {formatTime} from "../utils/common.js";
 
 export class TaskEdit extends AbstractComponent {
   constructor(task) {
