+++ src/components/task.js
@@ -1,6 +1,6 @@
-import {MONTH_NAMES} from "../const.js";
-import {formatTime} from "../utils.js";
 import {AbstractComponent} from "./abstract-component.js";
+import {MONTH_NAMES} from "../const.js";
+import {formatTime} from "../utils/common.js";
 
 export class Task extends AbstractComponent {
   constructor(task) {
