+++ src/utils/render.js
@@ -15,16 +15,6 @@ const render = (container, element, place) => {
   }
 };
 
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours() % 12);
-  const minutes = castTimeFormat(date.getMinutes());
-  return `${hours}:${minutes}`;
-};
-
 const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
@@ -33,4 +23,4 @@ const createElement = (template) => {
 };
 
 
-export {RenderPosition, render, formatTime, createElement};
+export {RenderPosition, render, createElement};
